[Edit] Shared 사용 이후 TDD 일부 주석 처리
